add simple_calendar fix twitter_user登録を登録した際、桁数エラーを修正

+++ db/migrate/20230110143031_create_twitter_users.rb
@@ -3,7 +3,7 @@ class CreateTwitterUsers < ActiveRecord::Migration[7.0]
     create_table :twitter_users, id: false do |t|
       t.string :name
       t.string :username
-      t.integer :user_id, null: false, primary_key: true
+      t.bigint :user_id, null: false, primary_key: true
 
       t.timestamp :created_at
     end
